Add EfficientNet code
Edited the runner to sub in the specified activation for single runs. Training started locally, so I think it works. I only swapped out our convolution for the first one in the pipeline; I don't understand the EfficientNet architecture well enough to know if/which other swish's should be swapped out.

Test plan: push and run on colab